% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab.R
\docType{data}
\name{ab}
\alias{ab}
\title{Official FWMIS Fish Status
Official Alberta Fisheries and Wildlife Management Information System (FWMIS) fish data.}
\format{
A tbl data frame that includes:
\describe{
\item{Species Common Name}{A character vector of the AB fish species common name.}
\item{Scientific Name}{A character vector of the AB fish species scientific name.}
\item{Species Code}{A character vector of the unique data centre fish code.}
}
}
\usage{
ab
}
\description{
Located under \verb{FWMIS fisheries loadforms}
}
\details{
The data is released under the \href{https://open.alberta.ca/licence}{Open Government Licence - Alberta}
}
\examples{
head(ab)
}
\references{
AB Fisheries and Wildlife Management Information System (FWMIS). 2021.
AB Environ. and Parks Edmonton, AB
Available: \url{https://www.alberta.ca/fisheries-loadforms.aspx}
(accessed Feb 9, 2021).
}
\seealso{
\link{freshwaterfish}
}
\keyword{datasets}
