% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbc-package.R
\docType{package}
\name{fishbc-package}
\alias{fishbc}
\alias{fishbc-package}
\title{fishbc: Fishes of British Columbia}
\description{
Provides raw and curated data on the codes,
    classification and conservation status of freshwater fishes in British
    Columbia. Marine fishes will be added in a future release.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/fishbc}
  \item Report bugs at \url{https://github.com/poissonconsulting/fishbc/issues}
}

}
\author{
\strong{Maintainer}: Evan Amies-Galonski \email{evan@poissonconsulting.ca} (\href{https://orcid.org/0000-0003-1096-2089}{ORCID})

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
}

Other contributors:
\itemize{
  \item Bronwen Lewis [contributor]
  \item Jesse Patterson [contributor]
  \item Gordon Oliphant [contributor]
  \item Seb Dalgarno [contributor]
  \item Poisson Consulting [copyright holder, funder]
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
