% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-name.R
\name{sbf_common_name}
\alias{sbf_common_name}
\title{Common Name}
\usage{
sbf_common_name(x)
}
\arguments{
\item{x}{A character vector of the provincial fish codes.}
}
\value{
A character vector of the corresponding common names.
}
\description{
Common Name
}
\examples{
sbf_common_name(c("AF", "WSG", NA, "AF", "NOTACODE"))
}
