% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab.R
\docType{data}
\name{ab}
\alias{ab}
\title{Official ACIMS Fish Status
Official Alberta Conservation Information Management System (ACIMS) fish data.}
\format{
A tbl data frame that includes:
\describe{
\item{Species Code}{A character vector of the unique data centre fish code.}
\item{AB List}{A character vector of the AB Conservation Listing.}
}
}
\usage{
ab
}
\description{
The search criteria for status was \verb{Select all}
and the class was \code{Fish}.
}
\details{
The data is released under the \href{https://open.alberta.ca/licence}{Open Government Licence - Alberta}
}
\examples{
head(ab)
}
\references{
AB Conservation Information Management System (ACIMS). 2021.
AB Fisheries and Wildlife Management Information System
AB Environ. and Parks Edmonton, AB
Available: \url{https://extranet.gov.ab.ca/env/wild-species-status/default.aspx}
(accessed Feb 9, 2021).
}
\seealso{
\link{freshwaterfish}
}
\keyword{datasets}
